IBM PROJECT/requirements.txt
flask==3.0.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
ibm-watsonx-ai==1.1.11
